Final Commit - 02/08/2021
Final Commit with the Programs, the Data and the Visualizations.
31/07/2021 - End of my Internship and my missions on the French Project «HyPe-13».
This project has been created to face the health crisis of the COVID-19 when the universities of France had to adapt their teaching methods due to the country lockdown with online or hybrid classes to ensure the educational continuity. Also, to make sure this project has the best chances of success, it has been created by a consortium of 12 french universities all over the country.